+++ src/main/java/br/ufes/sead/erp/financial/services/CourseService.java
@@ -1,12 +1,17 @@
 package br.ufes.sead.erp.financial.services;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import br.ufes.sead.erp.financial.entities.Course;
 import br.ufes.sead.erp.financial.repositories.CourseRepository;
+import br.ufes.sead.erp.financial.services.exceptions.DatabaseException;
+import br.ufes.sead.erp.financial.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class CourseService {
@@ -19,7 +24,8 @@ public class CourseService {
     }
 
     public Course findById(Long id) {
-        return courseRepository.findById(id).orElse(null);
+        Optional<Course> course = courseRepository.findById(id);
+        return course.orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
     public Course save(Course course) {
@@ -27,11 +33,37 @@ public class CourseService {
     }
 
     public void deleteById(Long id) {
-        courseRepository.deleteById(id);
+        try {
+            if (!courseRepository.existsById(id))
+                throw new ResourceNotFoundException(id);
+
+            courseRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public void delete(Course course) {
-        courseRepository.delete(course);
+        try {
+            if (!courseRepository.existsById(course.getId()))
+                throw new ResourceNotFoundException(course.getId());
+
+            courseRepository.delete(course);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
+    }
+
+    public Course update(Long id, Course course) {
+        try {
+            Course courseEntityReference = courseRepository.getReferenceById(id);
+
+            courseEntityReference.setName(course.getName());
+
+            return courseRepository.save(courseEntityReference);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException(id);
+        }
     }
 
 }
